refactor: move extension-related functionality in rich text representation model
