+++ src/modules/domain/rich-text/constants/index.ts
@@ -1,3 +1,2 @@
 export * from './blocks';
-export * from './representations';
 export * from './code-block-languages';
