+++ src/modules/domain/rich-text/models/index.ts
@@ -1,4 +1,5 @@
 export * from './document';
+export * from './representation';
 export * from './heading';
 export * from './link';
 export * from './document-asset';
