+++ src/modules/domain/rich-text/ports/document-analyzer.ts
@@ -1,8 +1,7 @@
 import * as Effect from 'effect/Effect';
 
-import { type RichTextRepresentation } from '../constants/representations';
 import { type DocumentAnalysisError, type RichTextLibError } from '../errors';
-import { type AssetDocRelPath } from '../models';
+import { type AssetDocRelPath, type RichTextRepresentation } from '../models';
 
 export type ExtractLocalAssetReferencesArgs = {
   representation: RichTextRepresentation;
